Checkpoint: File edits
Automatic checkpoint created by Claude Code
- Branch: v3
- Timestamp: 2026-01-07T04:41:53Z
- Changes: 1 file(s)

+++ v3/@claude-flow/embeddings/src/hyperbolic.ts
@@ -330,7 +330,7 @@ export function hyperbolicCentroid(
     for (let i = 0; i < dim; i++) {
       gradient[i] /= points.length;
     }
-    centroid = expMapAt(centroid, gradient, config);
+    centroid = expMapAt(centroid as Float32Array, gradient, config) as Float32Array;
   }
 
   return centroid;
